+++ aiku/aiku-main/src/main/java/aiku_main/controller/member/FirebaseController.java
@@ -1,4 +1,4 @@
-package aiku_main.controller;
+package aiku_main.controller.member;
 
 import aiku_main.dto.*;
 import aiku_main.service.member.MemberFirebaseService;
