+++ aiku/aiku-main/src/test/java/aiku_main/controller/TeamControllerTest.java
@@ -1,5 +1,6 @@
 package aiku_main.controller;
 
+import aiku_main.controller.team.TeamController;
 import aiku_main.dto.team.TeamAddDto;
 import aiku_main.service.team.TeamService;
 import com.fasterxml.jackson.databind.ObjectMapper;
